Update rust generated code to use flat_font::ufo namespace
