+++ ufoff/src/flatbuffers_generated/flat_font/ufo/woff_metadata_description_generated.rs
@@ -1,0 +1,113 @@
+// automatically generated by the FlatBuffers compiler, do not modify
+extern crate flatbuffers;
+use std::mem;
+use std::cmp::Ordering;
+use self::flatbuffers::{EndianScalar, Follow};
+use super::*;
+pub enum WoffMetadataDescriptionOffset {}
+#[derive(Copy, Clone, PartialEq)]
+
+pub struct WoffMetadataDescription<'a> {
+  pub _tab: flatbuffers::Table<'a>,
+}
+
+impl<'a> flatbuffers::Follow<'a> for WoffMetadataDescription<'a> {
+  type Inner = WoffMetadataDescription<'a>;
+  #[inline]
+  fn follow(buf: &'a [u8], loc: usize) -> Self::Inner {
+    Self { _tab: flatbuffers::Table { buf, loc } }
+  }
+}
+
+impl<'a> WoffMetadataDescription<'a> {
+  pub const VT_URL: flatbuffers::VOffsetT = 4;
+  pub const VT_TEXT: flatbuffers::VOffsetT = 6;
+
+  #[inline]
+  pub fn init_from_table(table: flatbuffers::Table<'a>) -> Self {
+    WoffMetadataDescription { _tab: table }
+  }
+  #[allow(unused_mut)]
+  pub fn create<'bldr: 'args, 'args: 'mut_bldr, 'mut_bldr>(
+    _fbb: &'mut_bldr mut flatbuffers::FlatBufferBuilder<'bldr>,
+    args: &'args WoffMetadataDescriptionArgs<'args>
+  ) -> flatbuffers::WIPOffset<WoffMetadataDescription<'bldr>> {
+    let mut builder = WoffMetadataDescriptionBuilder::new(_fbb);
+    if let Some(x) = args.text { builder.add_text(x); }
+    if let Some(x) = args.url { builder.add_url(x); }
+    builder.finish()
+  }
+
+
+  #[inline]
+  pub fn url(&self) -> Option<&'a str> {
+    self._tab.get::<flatbuffers::ForwardsUOffset<&str>>(WoffMetadataDescription::VT_URL, None)
+  }
+  #[inline]
+  pub fn text(&self) -> Option<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WoffMetadataText<'a>>>> {
+    self._tab.get::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WoffMetadataText>>>>(WoffMetadataDescription::VT_TEXT, None)
+  }
+}
+
+impl flatbuffers::Verifiable for WoffMetadataDescription<'_> {
+  #[inline]
+  fn run_verifier(
+    v: &mut flatbuffers::Verifier, pos: usize
+  ) -> Result<(), flatbuffers::InvalidFlatbuffer> {
+    use self::flatbuffers::Verifiable;
+    v.visit_table(pos)?
+     .visit_field::<flatbuffers::ForwardsUOffset<&str>>("url", Self::VT_URL, false)?
+     .visit_field::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'_, flatbuffers::ForwardsUOffset<WoffMetadataText>>>>("text", Self::VT_TEXT, false)?
+     .finish();
+    Ok(())
+  }
+}
+pub struct WoffMetadataDescriptionArgs<'a> {
+    pub url: Option<flatbuffers::WIPOffset<&'a str>>,
+    pub text: Option<flatbuffers::WIPOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WoffMetadataText<'a>>>>>,
+}
+impl<'a> Default for WoffMetadataDescriptionArgs<'a> {
+  #[inline]
+  fn default() -> Self {
+    WoffMetadataDescriptionArgs {
+      url: None,
+      text: None,
+    }
+  }
+}
+pub struct WoffMetadataDescriptionBuilder<'a: 'b, 'b> {
+  fbb_: &'b mut flatbuffers::FlatBufferBuilder<'a>,
+  start_: flatbuffers::WIPOffset<flatbuffers::TableUnfinishedWIPOffset>,
+}
+impl<'a: 'b, 'b> WoffMetadataDescriptionBuilder<'a, 'b> {
+  #[inline]
+  pub fn add_url(&mut self, url: flatbuffers::WIPOffset<&'b  str>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WoffMetadataDescription::VT_URL, url);
+  }
+  #[inline]
+  pub fn add_text(&mut self, text: flatbuffers::WIPOffset<flatbuffers::Vector<'b , flatbuffers::ForwardsUOffset<WoffMetadataText<'b >>>>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WoffMetadataDescription::VT_TEXT, text);
+  }
+  #[inline]
+  pub fn new(_fbb: &'b mut flatbuffers::FlatBufferBuilder<'a>) -> WoffMetadataDescriptionBuilder<'a, 'b> {
+    let start = _fbb.start_table();
+    WoffMetadataDescriptionBuilder {
+      fbb_: _fbb,
+      start_: start,
+    }
+  }
+  #[inline]
+  pub fn finish(self) -> flatbuffers::WIPOffset<WoffMetadataDescription<'a>> {
+    let o = self.fbb_.end_table(self.start_);
+    flatbuffers::WIPOffset::new(o.value())
+  }
+}
+
+impl std::fmt::Debug for WoffMetadataDescription<'_> {
+  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    let mut ds = f.debug_struct("WoffMetadataDescription");
+      ds.field("url", &self.url());
+      ds.field("text", &self.text());
+      ds.finish()
+  }
+}
