+++ ufoff/src/flatbuffers_generated/flat_font/ufo/woff_metadata_extension_generated.rs
@@ -1,0 +1,127 @@
+// automatically generated by the FlatBuffers compiler, do not modify
+extern crate flatbuffers;
+use std::mem;
+use std::cmp::Ordering;
+use self::flatbuffers::{EndianScalar, Follow};
+use super::*;
+pub enum WoffMetadataExtensionOffset {}
+#[derive(Copy, Clone, PartialEq)]
+
+pub struct WoffMetadataExtension<'a> {
+  pub _tab: flatbuffers::Table<'a>,
+}
+
+impl<'a> flatbuffers::Follow<'a> for WoffMetadataExtension<'a> {
+  type Inner = WoffMetadataExtension<'a>;
+  #[inline]
+  fn follow(buf: &'a [u8], loc: usize) -> Self::Inner {
+    Self { _tab: flatbuffers::Table { buf, loc } }
+  }
+}
+
+impl<'a> WoffMetadataExtension<'a> {
+  pub const VT_ID: flatbuffers::VOffsetT = 4;
+  pub const VT_NAMES: flatbuffers::VOffsetT = 6;
+  pub const VT_ITEMS: flatbuffers::VOffsetT = 8;
+
+  #[inline]
+  pub fn init_from_table(table: flatbuffers::Table<'a>) -> Self {
+    WoffMetadataExtension { _tab: table }
+  }
+  #[allow(unused_mut)]
+  pub fn create<'bldr: 'args, 'args: 'mut_bldr, 'mut_bldr>(
+    _fbb: &'mut_bldr mut flatbuffers::FlatBufferBuilder<'bldr>,
+    args: &'args WoffMetadataExtensionArgs<'args>
+  ) -> flatbuffers::WIPOffset<WoffMetadataExtension<'bldr>> {
+    let mut builder = WoffMetadataExtensionBuilder::new(_fbb);
+    if let Some(x) = args.items { builder.add_items(x); }
+    if let Some(x) = args.names { builder.add_names(x); }
+    if let Some(x) = args.id { builder.add_id(x); }
+    builder.finish()
+  }
+
+
+  #[inline]
+  pub fn id(&self) -> Option<&'a str> {
+    self._tab.get::<flatbuffers::ForwardsUOffset<&str>>(WoffMetadataExtension::VT_ID, None)
+  }
+  #[inline]
+  pub fn names(&self) -> Option<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WoffMetadataExtensionName<'a>>>> {
+    self._tab.get::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WoffMetadataExtensionName>>>>(WoffMetadataExtension::VT_NAMES, None)
+  }
+  #[inline]
+  pub fn items(&self) -> Option<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WoffMetadataExtensionItem<'a>>>> {
+    self._tab.get::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WoffMetadataExtensionItem>>>>(WoffMetadataExtension::VT_ITEMS, None)
+  }
+}
+
+impl flatbuffers::Verifiable for WoffMetadataExtension<'_> {
+  #[inline]
+  fn run_verifier(
+    v: &mut flatbuffers::Verifier, pos: usize
+  ) -> Result<(), flatbuffers::InvalidFlatbuffer> {
+    use self::flatbuffers::Verifiable;
+    v.visit_table(pos)?
+     .visit_field::<flatbuffers::ForwardsUOffset<&str>>("id", Self::VT_ID, false)?
+     .visit_field::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'_, flatbuffers::ForwardsUOffset<WoffMetadataExtensionName>>>>("names", Self::VT_NAMES, false)?
+     .visit_field::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'_, flatbuffers::ForwardsUOffset<WoffMetadataExtensionItem>>>>("items", Self::VT_ITEMS, false)?
+     .finish();
+    Ok(())
+  }
+}
+pub struct WoffMetadataExtensionArgs<'a> {
+    pub id: Option<flatbuffers::WIPOffset<&'a str>>,
+    pub names: Option<flatbuffers::WIPOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WoffMetadataExtensionName<'a>>>>>,
+    pub items: Option<flatbuffers::WIPOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WoffMetadataExtensionItem<'a>>>>>,
+}
+impl<'a> Default for WoffMetadataExtensionArgs<'a> {
+  #[inline]
+  fn default() -> Self {
+    WoffMetadataExtensionArgs {
+      id: None,
+      names: None,
+      items: None,
+    }
+  }
+}
+pub struct WoffMetadataExtensionBuilder<'a: 'b, 'b> {
+  fbb_: &'b mut flatbuffers::FlatBufferBuilder<'a>,
+  start_: flatbuffers::WIPOffset<flatbuffers::TableUnfinishedWIPOffset>,
+}
+impl<'a: 'b, 'b> WoffMetadataExtensionBuilder<'a, 'b> {
+  #[inline]
+  pub fn add_id(&mut self, id: flatbuffers::WIPOffset<&'b  str>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WoffMetadataExtension::VT_ID, id);
+  }
+  #[inline]
+  pub fn add_names(&mut self, names: flatbuffers::WIPOffset<flatbuffers::Vector<'b , flatbuffers::ForwardsUOffset<WoffMetadataExtensionName<'b >>>>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WoffMetadataExtension::VT_NAMES, names);
+  }
+  #[inline]
+  pub fn add_items(&mut self, items: flatbuffers::WIPOffset<flatbuffers::Vector<'b , flatbuffers::ForwardsUOffset<WoffMetadataExtensionItem<'b >>>>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WoffMetadataExtension::VT_ITEMS, items);
+  }
+  #[inline]
+  pub fn new(_fbb: &'b mut flatbuffers::FlatBufferBuilder<'a>) -> WoffMetadataExtensionBuilder<'a, 'b> {
+    let start = _fbb.start_table();
+    WoffMetadataExtensionBuilder {
+      fbb_: _fbb,
+      start_: start,
+    }
+  }
+  #[inline]
+  pub fn finish(self) -> flatbuffers::WIPOffset<WoffMetadataExtension<'a>> {
+    let o = self.fbb_.end_table(self.start_);
+    flatbuffers::WIPOffset::new(o.value())
+  }
+}
+
+impl std::fmt::Debug for WoffMetadataExtension<'_> {
+  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    let mut ds = f.debug_struct("WoffMetadataExtension");
+      ds.field("id", &self.id());
+      ds.field("names", &self.names());
+      ds.field("items", &self.items());
+      ds.finish()
+  }
+}
