+++ ufoff/src/flatbuffers_generated/flat_font/ufo/woff_metadata_licensee_generated.rs
@@ -1,0 +1,128 @@
+// automatically generated by the FlatBuffers compiler, do not modify
+extern crate flatbuffers;
+use std::mem;
+use std::cmp::Ordering;
+use self::flatbuffers::{EndianScalar, Follow};
+use super::*;
+pub enum WoffMetadataLicenseeOffset {}
+#[derive(Copy, Clone, PartialEq)]
+
+pub struct WoffMetadataLicensee<'a> {
+  pub _tab: flatbuffers::Table<'a>,
+}
+
+impl<'a> flatbuffers::Follow<'a> for WoffMetadataLicensee<'a> {
+  type Inner = WoffMetadataLicensee<'a>;
+  #[inline]
+  fn follow(buf: &'a [u8], loc: usize) -> Self::Inner {
+    Self { _tab: flatbuffers::Table { buf, loc } }
+  }
+}
+
+impl<'a> WoffMetadataLicensee<'a> {
+  pub const VT_NAME: flatbuffers::VOffsetT = 4;
+  pub const VT_DIR: flatbuffers::VOffsetT = 6;
+  pub const VT_CLASS: flatbuffers::VOffsetT = 8;
+
+  #[inline]
+  pub fn init_from_table(table: flatbuffers::Table<'a>) -> Self {
+    WoffMetadataLicensee { _tab: table }
+  }
+  #[allow(unused_mut)]
+  pub fn create<'bldr: 'args, 'args: 'mut_bldr, 'mut_bldr>(
+    _fbb: &'mut_bldr mut flatbuffers::FlatBufferBuilder<'bldr>,
+    args: &'args WoffMetadataLicenseeArgs<'args>
+  ) -> flatbuffers::WIPOffset<WoffMetadataLicensee<'bldr>> {
+    let mut builder = WoffMetadataLicenseeBuilder::new(_fbb);
+    if let Some(x) = args.class { builder.add_class(x); }
+    if let Some(x) = args.dir { builder.add_dir(x); }
+    if let Some(x) = args.name { builder.add_name(x); }
+    builder.finish()
+  }
+
+
+  #[inline]
+  pub fn name(&self) -> &'a str {
+    self._tab.get::<flatbuffers::ForwardsUOffset<&str>>(WoffMetadataLicensee::VT_NAME, None).unwrap()
+  }
+  #[inline]
+  pub fn dir(&self) -> Option<&'a str> {
+    self._tab.get::<flatbuffers::ForwardsUOffset<&str>>(WoffMetadataLicensee::VT_DIR, None)
+  }
+  #[inline]
+  pub fn class(&self) -> Option<&'a str> {
+    self._tab.get::<flatbuffers::ForwardsUOffset<&str>>(WoffMetadataLicensee::VT_CLASS, None)
+  }
+}
+
+impl flatbuffers::Verifiable for WoffMetadataLicensee<'_> {
+  #[inline]
+  fn run_verifier(
+    v: &mut flatbuffers::Verifier, pos: usize
+  ) -> Result<(), flatbuffers::InvalidFlatbuffer> {
+    use self::flatbuffers::Verifiable;
+    v.visit_table(pos)?
+     .visit_field::<flatbuffers::ForwardsUOffset<&str>>("name", Self::VT_NAME, true)?
+     .visit_field::<flatbuffers::ForwardsUOffset<&str>>("dir", Self::VT_DIR, false)?
+     .visit_field::<flatbuffers::ForwardsUOffset<&str>>("class", Self::VT_CLASS, false)?
+     .finish();
+    Ok(())
+  }
+}
+pub struct WoffMetadataLicenseeArgs<'a> {
+    pub name: Option<flatbuffers::WIPOffset<&'a str>>,
+    pub dir: Option<flatbuffers::WIPOffset<&'a str>>,
+    pub class: Option<flatbuffers::WIPOffset<&'a str>>,
+}
+impl<'a> Default for WoffMetadataLicenseeArgs<'a> {
+  #[inline]
+  fn default() -> Self {
+    WoffMetadataLicenseeArgs {
+      name: None, // required field
+      dir: None,
+      class: None,
+    }
+  }
+}
+pub struct WoffMetadataLicenseeBuilder<'a: 'b, 'b> {
+  fbb_: &'b mut flatbuffers::FlatBufferBuilder<'a>,
+  start_: flatbuffers::WIPOffset<flatbuffers::TableUnfinishedWIPOffset>,
+}
+impl<'a: 'b, 'b> WoffMetadataLicenseeBuilder<'a, 'b> {
+  #[inline]
+  pub fn add_name(&mut self, name: flatbuffers::WIPOffset<&'b  str>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WoffMetadataLicensee::VT_NAME, name);
+  }
+  #[inline]
+  pub fn add_dir(&mut self, dir: flatbuffers::WIPOffset<&'b  str>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WoffMetadataLicensee::VT_DIR, dir);
+  }
+  #[inline]
+  pub fn add_class(&mut self, class: flatbuffers::WIPOffset<&'b  str>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WoffMetadataLicensee::VT_CLASS, class);
+  }
+  #[inline]
+  pub fn new(_fbb: &'b mut flatbuffers::FlatBufferBuilder<'a>) -> WoffMetadataLicenseeBuilder<'a, 'b> {
+    let start = _fbb.start_table();
+    WoffMetadataLicenseeBuilder {
+      fbb_: _fbb,
+      start_: start,
+    }
+  }
+  #[inline]
+  pub fn finish(self) -> flatbuffers::WIPOffset<WoffMetadataLicensee<'a>> {
+    let o = self.fbb_.end_table(self.start_);
+    self.fbb_.required(o, WoffMetadataLicensee::VT_NAME,"name");
+    flatbuffers::WIPOffset::new(o.value())
+  }
+}
+
+impl std::fmt::Debug for WoffMetadataLicensee<'_> {
+  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    let mut ds = f.debug_struct("WoffMetadataLicensee");
+      ds.field("name", &self.name());
+      ds.field("dir", &self.dir());
+      ds.field("class", &self.class());
+      ds.finish()
+  }
+}
