+++ ufoff/src/flatbuffers_generated/flat_font/ufo/woff_metadata_vendor_generated.rs
@@ -1,0 +1,142 @@
+// automatically generated by the FlatBuffers compiler, do not modify
+extern crate flatbuffers;
+use std::mem;
+use std::cmp::Ordering;
+use self::flatbuffers::{EndianScalar, Follow};
+use super::*;
+pub enum WoffMetadataVendorOffset {}
+#[derive(Copy, Clone, PartialEq)]
+
+pub struct WoffMetadataVendor<'a> {
+  pub _tab: flatbuffers::Table<'a>,
+}
+
+impl<'a> flatbuffers::Follow<'a> for WoffMetadataVendor<'a> {
+  type Inner = WoffMetadataVendor<'a>;
+  #[inline]
+  fn follow(buf: &'a [u8], loc: usize) -> Self::Inner {
+    Self { _tab: flatbuffers::Table { buf, loc } }
+  }
+}
+
+impl<'a> WoffMetadataVendor<'a> {
+  pub const VT_NAME: flatbuffers::VOffsetT = 4;
+  pub const VT_URL: flatbuffers::VOffsetT = 6;
+  pub const VT_DIR: flatbuffers::VOffsetT = 8;
+  pub const VT_CLASS: flatbuffers::VOffsetT = 10;
+
+  #[inline]
+  pub fn init_from_table(table: flatbuffers::Table<'a>) -> Self {
+    WoffMetadataVendor { _tab: table }
+  }
+  #[allow(unused_mut)]
+  pub fn create<'bldr: 'args, 'args: 'mut_bldr, 'mut_bldr>(
+    _fbb: &'mut_bldr mut flatbuffers::FlatBufferBuilder<'bldr>,
+    args: &'args WoffMetadataVendorArgs<'args>
+  ) -> flatbuffers::WIPOffset<WoffMetadataVendor<'bldr>> {
+    let mut builder = WoffMetadataVendorBuilder::new(_fbb);
+    if let Some(x) = args.class { builder.add_class(x); }
+    if let Some(x) = args.dir { builder.add_dir(x); }
+    if let Some(x) = args.url { builder.add_url(x); }
+    if let Some(x) = args.name { builder.add_name(x); }
+    builder.finish()
+  }
+
+
+  #[inline]
+  pub fn name(&self) -> &'a str {
+    self._tab.get::<flatbuffers::ForwardsUOffset<&str>>(WoffMetadataVendor::VT_NAME, None).unwrap()
+  }
+  #[inline]
+  pub fn url(&self) -> Option<&'a str> {
+    self._tab.get::<flatbuffers::ForwardsUOffset<&str>>(WoffMetadataVendor::VT_URL, None)
+  }
+  #[inline]
+  pub fn dir(&self) -> Option<&'a str> {
+    self._tab.get::<flatbuffers::ForwardsUOffset<&str>>(WoffMetadataVendor::VT_DIR, None)
+  }
+  #[inline]
+  pub fn class(&self) -> Option<&'a str> {
+    self._tab.get::<flatbuffers::ForwardsUOffset<&str>>(WoffMetadataVendor::VT_CLASS, None)
+  }
+}
+
+impl flatbuffers::Verifiable for WoffMetadataVendor<'_> {
+  #[inline]
+  fn run_verifier(
+    v: &mut flatbuffers::Verifier, pos: usize
+  ) -> Result<(), flatbuffers::InvalidFlatbuffer> {
+    use self::flatbuffers::Verifiable;
+    v.visit_table(pos)?
+     .visit_field::<flatbuffers::ForwardsUOffset<&str>>("name", Self::VT_NAME, true)?
+     .visit_field::<flatbuffers::ForwardsUOffset<&str>>("url", Self::VT_URL, false)?
+     .visit_field::<flatbuffers::ForwardsUOffset<&str>>("dir", Self::VT_DIR, false)?
+     .visit_field::<flatbuffers::ForwardsUOffset<&str>>("class", Self::VT_CLASS, false)?
+     .finish();
+    Ok(())
+  }
+}
+pub struct WoffMetadataVendorArgs<'a> {
+    pub name: Option<flatbuffers::WIPOffset<&'a str>>,
+    pub url: Option<flatbuffers::WIPOffset<&'a str>>,
+    pub dir: Option<flatbuffers::WIPOffset<&'a str>>,
+    pub class: Option<flatbuffers::WIPOffset<&'a str>>,
+}
+impl<'a> Default for WoffMetadataVendorArgs<'a> {
+  #[inline]
+  fn default() -> Self {
+    WoffMetadataVendorArgs {
+      name: None, // required field
+      url: None,
+      dir: None,
+      class: None,
+    }
+  }
+}
+pub struct WoffMetadataVendorBuilder<'a: 'b, 'b> {
+  fbb_: &'b mut flatbuffers::FlatBufferBuilder<'a>,
+  start_: flatbuffers::WIPOffset<flatbuffers::TableUnfinishedWIPOffset>,
+}
+impl<'a: 'b, 'b> WoffMetadataVendorBuilder<'a, 'b> {
+  #[inline]
+  pub fn add_name(&mut self, name: flatbuffers::WIPOffset<&'b  str>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WoffMetadataVendor::VT_NAME, name);
+  }
+  #[inline]
+  pub fn add_url(&mut self, url: flatbuffers::WIPOffset<&'b  str>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WoffMetadataVendor::VT_URL, url);
+  }
+  #[inline]
+  pub fn add_dir(&mut self, dir: flatbuffers::WIPOffset<&'b  str>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WoffMetadataVendor::VT_DIR, dir);
+  }
+  #[inline]
+  pub fn add_class(&mut self, class: flatbuffers::WIPOffset<&'b  str>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WoffMetadataVendor::VT_CLASS, class);
+  }
+  #[inline]
+  pub fn new(_fbb: &'b mut flatbuffers::FlatBufferBuilder<'a>) -> WoffMetadataVendorBuilder<'a, 'b> {
+    let start = _fbb.start_table();
+    WoffMetadataVendorBuilder {
+      fbb_: _fbb,
+      start_: start,
+    }
+  }
+  #[inline]
+  pub fn finish(self) -> flatbuffers::WIPOffset<WoffMetadataVendor<'a>> {
+    let o = self.fbb_.end_table(self.start_);
+    self.fbb_.required(o, WoffMetadataVendor::VT_NAME,"name");
+    flatbuffers::WIPOffset::new(o.value())
+  }
+}
+
+impl std::fmt::Debug for WoffMetadataVendor<'_> {
+  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    let mut ds = f.debug_struct("WoffMetadataVendor");
+      ds.field("name", &self.name());
+      ds.field("url", &self.url());
+      ds.field("dir", &self.dir());
+      ds.field("class", &self.class());
+      ds.finish()
+  }
+}
