+++ ufoff/src/flatbuffers_generated/mod.rs
@@ -1,0 +1,45 @@
+// Automatically generated by the Flatbuffers compiler. Do not modify.
+pub mod flat_font {
+  use super::*;
+  pub mod ufo {
+    use super::*;
+    mod guideline_generated;
+    pub use self::guideline_generated::*;
+    mod transform_generated;
+    pub use self::transform_generated::*;
+    mod font_info_generated;
+    pub use self::font_info_generated::*;
+    mod open_type_name_record_generated;
+    pub use self::open_type_name_record_generated::*;
+    mod open_type_gasp_range_record_generated;
+    pub use self::open_type_gasp_range_record_generated::*;
+    mod woff_metadata_unique_id_generated;
+    pub use self::woff_metadata_unique_id_generated::*;
+    mod woff_metadata_vendor_generated;
+    pub use self::woff_metadata_vendor_generated::*;
+    mod woff_metadata_credits_generated;
+    pub use self::woff_metadata_credits_generated::*;
+    mod woff_metadata_credit_generated;
+    pub use self::woff_metadata_credit_generated::*;
+    mod woff_metadata_description_generated;
+    pub use self::woff_metadata_description_generated::*;
+    mod woff_metadata_license_generated;
+    pub use self::woff_metadata_license_generated::*;
+    mod woff_metadata_copyright_generated;
+    pub use self::woff_metadata_copyright_generated::*;
+    mod woff_metadata_trademark_generated;
+    pub use self::woff_metadata_trademark_generated::*;
+    mod woff_metadata_licensee_generated;
+    pub use self::woff_metadata_licensee_generated::*;
+    mod woff_metadata_text_generated;
+    pub use self::woff_metadata_text_generated::*;
+    mod woff_metadata_extension_generated;
+    pub use self::woff_metadata_extension_generated::*;
+    mod woff_metadata_extension_item_generated;
+    pub use self::woff_metadata_extension_item_generated::*;
+    mod woff_metadata_extension_name_generated;
+    pub use self::woff_metadata_extension_name_generated::*;
+    mod woff_metadata_extension_value_generated;
+    pub use self::woff_metadata_extension_value_generated::*;
+  } // ufo
+} // flat_font
